refactor(agent): isolate canonical history validation
